backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
paho-mqtt==1.6.1
PyMySQL==1.1.1
cryptography==42.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
eventlet==0.36.0
gevent==24.2.1
simple-websocket==1.0.0
